scripts/utils.py
import torch
import torchvision
from torchvision import transforms
from torch.utils.data import DataLoader
import os
import cv2
import numpy as np

def load_data(image_dir, label_dir):
    # Placeholder for loading data
    # Implement your data loading logic here
    pass

def train_model(data, device):
    # Placeholder for training model
    # Implement your training logic here
    model = torchvision.models.detection.fasterrcnn_resnet50_fpn(pretrained=True)
    model.to(device)
    model.train()
    # Add training loop here
    return model

def load_model(model_path, device):
    model = torchvision.models.detection.fasterrcnn_resnet50_fpn(pretrained=False)
    model.load_state_dict(torch.load(model_path))
    model.to(device)
    model.eval()
    return model

def detect_objects(model, image_dir, device):
    transform = transforms.Compose([
        transforms.ToTensor(),
    ])
    for img_name in os.listdir(image_dir):
        img_path = os.path.join(image_dir, img_name)
        image = cv2.imread(img_path)
        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
        image = transform(image).unsqueeze(0).to(device)
        with torch.no_grad():
            outputs = model(image)
        # Process outputs
        print(outputs) 