requirements.txt
ultralytics
torch
torchvision
numpy
opencv-python
matplotlib
tabulate
pandas
seaborn
scikit-learn
pyyaml